Re-remove memory mapping functions
There are three reasons why they are removed:

First, because the mapping functions are no longer needed by QMKL, since
memory allocation is now done by VCSM, which is one of the Raspberry Pi
Foundation's official driver.

Second, the mapping function always mapped memory as non-cached, which
confused users.  FYI, to enable caching of mapped memory with /dev/mem,
you need to memblock_add() in the Linux kernel and not to specify O_SYNC
flag, which cannot be achieved without a help of your own kernel driver.

Third, removing the mapping function makes QMKL only need to open
/dev/vcio and /dev/vcsm, which owner is root.video and their permission
is rw-rw----.  So now if you belong to video group, you do NOT need to
sudo to execute program on QPU!